+++ tests/resources/tf/container.tf
@@ -1,8 +1,8 @@
 resource "aws_vpc" "parent" {
-  name="parent"
+    name = "parent"
 }
 
 resource "aws_internet_gateway" "child" {
-  name="child"
-  vpc_id="parent"
+    name = "child"
+    vpc_id = "parent"
 }